工单操作公共方法

+++ Code/web/dms_repair/com.hsweb.repair/src/webcontent/js/RepairBusiness/Reception/carWashBill.js
@@ -265,14 +265,39 @@ $(document).ready(function ()
         var rowIndex = e.rowIndex;
 
         switch (e.field) {
-            case "packageOptBtn":
-                var s = '<a class="optbtn" href="javascript:newRow()">修改</a>'
-                        + ' <a class="optbtn" href="javascript:deletePackRow(\'' + uid + '\')">删除</a>';
-
-                if (grid.isEditingRow(record)) {
-                    s = '<a class="optbtn" href="javascript:newRow()">确定</a>'
-                        + ' <a class="optbtn" href="javascript:deletePackRow(\'' + uid + '\')">删除</a>';
+            case "prdtName":
+                var cardDetailId = record.cardDetailId||0;
+                if(cardDetailId>0){
+                    e.cellHtml = e.value + "(预存)";
                 }
+                break;
+            case "serviceTypeId":
+                var type = record.type||0;
+                if(type>1){
+                    e.cellHtml = "--";
+                }else{
+                    e.cellHtml = servieTypeHash[e.value].name;
+                }
+                break;
+            case "packageOptBtn":
+                var pid = record.pid||0;
+
+                if(pid == 0){
+                    var s = '<a class="optbtn" href="javascript:editRpsPackage(\'' + uid + '\')">修改</a>'
+                    + ' <a class="optbtn" href="javascript:deletePackRow(\'' + uid + '\')">删除</a>';
+
+                    if (grid.isEditingRow(record)) {
+                        s = '<a class="optbtn" href="javascript:updateRpsPackage(\'' + uid + '\')">确定</a>'
+                            + ' <a class="optbtn" href="javascript:deletePackRow(\'' + uid + '\')">删除</a>';
+                    }
+                }else{
+                    var s = '<a class="optbtn" href="javascript:editRpsPackage(\'' + uid + '\')">修改</a>';
+
+                    if (grid.isEditingRow(record)) {
+                        s = '<a class="optbtn" href="javascript:updateRpsPackage(\'' + uid + '\')">确定</a>';
+                    }
+                }
+                
                 e.cellHtml = s;
                  //'<span class="fa fa-close fa-lg" onClick="javascript:deletePart()" title="删除行">&nbsp;&nbsp;&nbsp;&nbsp;</span>';
                            // '<span class="fa fa-plus" onClick="javascript:addPackNewRow()" title="添加行">&nbsp;&nbsp;</span>' +
@@ -294,6 +319,38 @@ $(document).ready(function ()
                 break;
         }
     });
+    rpsPackageGrid.on("cellbeginedit",function(e){
+        var field=e.field; 
+        var editor = e.editor;
+        var row = e.row;
+        var column = e.column;
+        var editor = e.editor;
+
+        if(field == 'serviceTypeId'){
+            if(row.type > 1) {
+                e.cancel = true;
+            }
+        }
+        if(field == 'subtotal' || field == 'rate'){
+            if(row.type > 1) {
+                e.cancel = true;
+            }else{
+                if(row.cardDetailId > 0){
+                    e.cancel = true;
+                }
+            }
+        }
+        if(field == 'workers'){
+            if(row.type != 2){
+                e.cancel = true;
+            }
+        }
+        if(field == 'saleMan'){
+            if(row.type != 1 || row.cardDetailId > 0){
+                e.cancel = true;
+            }
+        }
+    });
     rpsItemGrid.on("drawcell", function (e) {
         var grid = e.sender;
         var record = e.record;
@@ -301,12 +358,18 @@ $(document).ready(function ()
         var rowIndex = e.rowIndex;
 
         switch (e.field) {
+            case "itemName":
+                var cardDetailId = record.cardDetailId||0;
+                if(cardDetailId>0){
+                    e.cellHtml = e.value + "(预存)";
+                }
+                break;
             case "itemOptBtn":
-                var s = '<a class="optbtn" href="javascript:newRow()">修改</a>'
+                var s = '<a class="optbtn" href="javascript:editRpsItem(\'' + uid + '\')">修改</a>'
                         + ' <a class="optbtn" href="javascript:deleteItemRow(\'' + uid + '\')">删除</a>';
 
                 if (grid.isEditingRow(record)) {
-                    s = '<a class="optbtn" href="javascript:newRow()">确定</a>'
+                    s = '<a class="optbtn" href="javascript:updateRpsItem(\'' + uid + '\')">确定</a>'
                         + ' <a class="optbtn" href="javascript:deleteItemRow(\'' + uid + '\')">删除</a>';
                 }
 
@@ -326,12 +389,18 @@ $(document).ready(function ()
         var rowIndex = e.rowIndex;
 
         switch (e.field) {
+            case "partName":
+                var cardDetailId = record.cardDetailId||0;
+                if(cardDetailId>0){
+                    e.cellHtml = e.value + "(预存)";
+                }
+                break;
             case "partOptBtn":
-                var s = '<a class="optbtn" href="javascript:newRow()">修改</a>'
+                var s = '<a class="optbtn" href="javascript:editRpsPart(\'' + uid + '\')">修改</a>'
                         + ' <a class="optbtn" href="javascript:deletePartRow(\'' + uid + '\')">删除</a>';
 
                 if (grid.isEditingRow(record)) {
-                    s = '<a class="optbtn" href="javascript:newRow()">确定</a>'
+                    s = '<a class="optbtn" href="javascript:updateRpsPart(\'' + uid + '\')">确定</a>'
                         + ' <a class="optbtn" href="javascript:deletePartRow(\'' + uid + '\')">删除</a>';
                 }
                 e.cellHtml = s;
@@ -881,26 +950,73 @@ function addPartNewRow(){
 }
 function deletePackRow(row_uid){
     var data = rpsPackageGrid.getData();
+    var row = rpsPackageGrid.getRowByUID(row_uid);
     if(data && data.length==1){
-        var row = rpsPackageGrid.getSelected();
-        rpsPackageGrid.removeRow(row);
-        var newRow = {};
-        rpsPackageGrid.addRow(newRow);
-    }else{
-        var row = rpsPackageGrid.getSelected();
-        rpsPackageGrid.removeRow(row);
+        row = data[0];
     }
+    var pkg = {
+        serviceId:row.serviceId,
+        id:row.id,
+        cardDetailId:row.cardDetailId||0
+    };
+    var params = {
+        type:"delete",
+        interType:"package",
+        data:{
+            pkg: pkg
+        }
+    };
+    svrCRUD(params,function(text){
+        var errCode = text.errCode||"";
+        var errMsg = text.errMsg||"";
+        if(errCode == 'S'){   
+            var rows = rpsPackageGrid.findRows(function(row){
+                if(row.prdtId == row.prdtId || row.pid == row.prdtID){
+                    return true;
+                }
+            });
+
+            rpsPackageGrid.removeRows(rows);
+        }else{
+            showMsg(errMsg||"删除套餐信息失败!","W");
+            return;
+        }
+    });
 }
 function deleteItemRow(row_uid){
     var data = rpsItemGrid.getData();
     var row = rpsItemGrid.getRowByUID(row_uid);
     if(data && data.length==1){
-        rpsItemGrid.removeRow(data[0]);
-        var newRow = {};
-        rpsItemGrid.addRow(newRow);
-    }else{
-        rpsItemGrid.removeRow(row);
+        row = data[0];
     }
+    var item = {
+        serviceId:row.serviceId,
+        id:row.id,
+        cardDetailId:row.cardDetailId||0
+    };
+    var params = {
+        type:"delete",
+        interType:"item",
+        data:{
+            item: item
+        }
+    };
+    svrCRUD(params,function(text){
+        var errCode = text.errCode||"";
+        var errMsg = text.errMsg||"";
+        if(errCode == 'S'){   
+            if(data && data.length==1){
+                rpsItemGrid.removeRow(data[0]);
+                //var newRow = {};
+                //rpsPartGrid.addRow(newRow);
+            }else{
+                rpsItemGrid.removeRow(row);
+            }
+        }else{
+            showMsg(errMsg||"删除工时信息失败!","W");
+            return;
+        }
+    });
 }
 function deletePartRow(row_uid){
     var data = rpsPartGrid.getData();
@@ -922,16 +1038,17 @@ function deletePartRow(row_uid){
     };
     svrCRUD(params,function(text){
         var errCode = text.errCode||"";
+        var errMsg = text.errMsg||"";
         if(errCode == 'S'){   
             if(data && data.length==1){
                 rpsPartGrid.removeRow(data[0]);
-                var newRow = {};
-                rpsPartGrid.addRow(newRow);
+                //var newRow = {};
+                //rpsPartGrid.addRow(newRow);
             }else{
                 rpsPartGrid.removeRow(row);
             }
         }else{
-            showMsg("删除配件信息失败!","W");
+            showMsg(errMsg||"删除配件信息失败!","W");
             return;
         }
     });
@@ -1062,7 +1179,7 @@ function onworkerChanged(e){
     if(!rows || rows.length==0){
         workerIds = "";
     }else{
-        for(var i=i; i<rows.length; i++){
+        for(var i=0; i<rows.length; i++){
             var row = rows[i];
             var empId = row.empId;
             workerIdList.push(empId);
@@ -1076,7 +1193,70 @@ function onworkerChanged(e){
     }
     var row = rpsPackageGrid.getSelected();
     var newRow = {workerIds:workerIds};
-    rpsPackageGrid.updateRow(row, newRow);
+    __workerIds = workerIds;
+}
+function onitemworkerChanged(e){
+    var obj = e.sender;
+    var rows = e.selecteds;
+    var workerIds = "";
+    var workerIdList = [];
+    if(!rows || rows.length==0){
+        workerIds = "";
+    }else{
+        for(var i=0; i<rows.length; i++){
+            var row = rows[i];
+            var empId = row.empId;
+            workerIdList.push(empId);
+        }
+
+        if(workerIdList&&workerIdList.length>0){
+            workerIds = workerIdList.join(",");
+        }else{
+            workerIds = "";
+        }
+    }
+    var row = rpsItemGrid.getSelected();
+    var newRow = {workerIds:workerIds};
+    rpsItemGrid.updateRow(row, newRow);
+}
+function onsalemanChanged(e){
+    var obj = e.sender;
+    var row = e.selected;
+    var saleManId = 0;
+    if(!row){
+        saleManId = 0;
+    }else{
+        saleManId = row.empId;
+    }
+    var row = rpsPackageGrid.getSelected();
+    __saleManId = saleManId;
+    //rpsPackageGrid.updateRow(row, newRow);
+}
+function onitemsalemanChanged(e){
+    var obj = e.sender;
+    var row = e.selected;
+    var saleManId = 0;
+    if(!row){
+        saleManId = 0;
+    }else{
+        saleManId = row.empId;
+    }
+    var row = rpsItemGrid.getSelected();
+    var newRow = {saleManId:saleManId};
+    rpsItemGrid.updateRow(row, newRow);
+}
+function onpartsalemanChanged(e){
+    var obj = e.sender;
+    var row = e.selected;
+    var saleManId = 0;
+    if(!row){
+        saleManId = 0;
+    }else{
+        saleManId = row.empId;
+    }
+    var row = rpsPartGrid.getSelected();
+    var newRow = {saleManId:saleManId};
+    rpsPartGrid.updateRow(row, newRow);
 }
 function addCardTimesToBill(){
     var main = billForm.getData();
@@ -1119,7 +1299,8 @@ function addCardTimesToBill(){
             var insPart = {
                 serviceId:main.id||0,
                 partId:row.prdtId,
-                cardDetailId:row.id||0
+                cardDetailId:row.id||0,
+                partCode:row.prdtCode
             };
             data.insPart = insPart;
             data.serviceId = main.id||0;
@@ -1131,6 +1312,7 @@ function addCardTimesToBill(){
         };
         svrCRUD(params,function(text){
             var errCode = text.errCode||"";
+            var errMsg = text.errMsg||"";
             if(errCode == 'S'){
                 //showMsg("添加次卡信息成功!","W");
                 //根据工单ID查询套餐,隐藏次卡信息
@@ -1160,7 +1342,7 @@ function addCardTimesToBill(){
                     }
                 }, function(){});
             }else{
-                showMsg("添加次卡信息失败!","W");
+                showMsg(errMsg||"添加预存信息失败!","W");
                 return;
             }
         });
@@ -1177,6 +1359,7 @@ function loadDetail(p1, p2, p3){
             if(errCode == "S"){
                 rpsPackageGrid.clearRows();
                 rpsPackageGrid.addRows(data);
+                rpsPackageGrid.accept();
             }
         }, function(){});
     }
@@ -1187,6 +1370,7 @@ function loadDetail(p1, p2, p3){
             if(errCode == "S"){
                 rpsItemGrid.clearRows();
                 rpsItemGrid.addRows(data);
+                rpsItemGrid.accept();
             }
         }, function(){});
     }
@@ -1197,8 +1381,97 @@ function loadDetail(p1, p2, p3){
             if(errCode == "S"){
                 rpsPartGrid.clearRows();
                 rpsPartGrid.addRows(data);
+                rpsPartGrid.accept();
             }
         }, function(){});
     }
 
+}
+var __workerIds="";
+var __saleManId="";
+function editRpsPackage(row_uid){
+    var row = rpsPackageGrid.getRowByUID(row_uid);
+    if (row) {
+        __workerIds = "";
+        __saleManId = "";
+        rpsPackageGrid.cancelEdit();
+        rpsPackageGrid.beginEditRow(row);
+    }
+}
+function updateRpsPackage(row_uid){
+    var rowc = rpsPackageGrid.getRowByUID(row_uid);
+    if (rowc) {
+        rpsPackageGrid.commitEdit();
+        var rows = rpsPackageGrid.getChanges();
+
+        var isSubtotalModify = 0;
+        if(rows && rows.length>0){
+            var row = rows[0];
+            if(row.type == 3){
+                rpsPackageGrid.accept();
+                return;
+            }
+            var serviceId = row.serviceId||0;
+            var cardDetailId = row.cardDetailId||0;
+            var pkg = {
+                serviceId:row.serviceId
+            };
+            var itemList = [];
+            if(row.type == 1){
+                pkg.id = row.id||0;
+                if(cardDetailId > 0){
+                    pkg.serviceTypeId = row.serviceTypeId;
+                }else{
+                    pkg.subtotal = row.subtotal
+                    pkg.serviceTypeId = row.serviceTypeId;
+                    if(__saleManId){
+                        pkg.saleMan = row.saleMan;
+                        pkg.saleManId = __saleManId;
+                    }
+                    isSubtotalModify = 1;
+                }
+            }else{
+                pkg.id = row.billPackageId||0;
+                if(__workerIds){
+                    var item = {
+                        id: row.id,
+                        serviceId: row.serviceId,
+                        workerIds:__workerIds,
+                        workers:row.workers
+                    }
+                    itemList.push(item);
+                }
+            }
+            var params = {
+                type:"update",
+                interType:"package",
+                data:{
+                    pkg: pkg,
+                    itemList : itemList
+                }
+            };
+
+            svrCRUD(params,function(text){
+                var errCode = text.errCode||"";
+                var errMsg = text.errMsg||"";
+                if(errCode == 'S'){   
+                    __workerIds = "";
+                    __saleManId = "";
+                    rpsPackageGrid.accept();
+                    if(isSubtotalModify == 1){
+                        var p1 = {
+                            interType: "package",
+                            data:{
+                                serviceId: serviceId
+                            }
+                        }
+                        loadDetail(p1, {}, {});
+                    }
+                }else{
+                    showMsg(errMsg||"修改数据失败!","W");
+                    return;
+                }
+            });
+        }
+    }
 }
